Remove blockFactory.
