+++ node/impls.go
@@ -18,16 +18,12 @@ package node
 
 import (
 	"context"
-	"time"
 
 	"github.com/algorand/go-algorand/agreement"
 	"github.com/algorand/go-algorand/catchup"
 	"github.com/algorand/go-algorand/data"
-	"github.com/algorand/go-algorand/data/basics"
 	"github.com/algorand/go-algorand/data/bookkeeping"
-	"github.com/algorand/go-algorand/data/committee"
 	"github.com/algorand/go-algorand/data/pools"
-	"github.com/algorand/go-algorand/ledger"
 	"github.com/algorand/go-algorand/logging"
 	"github.com/algorand/go-algorand/network"
 	"github.com/algorand/go-algorand/util/execpool"
@@ -63,36 +59,6 @@ func (i blockValidatorImpl) Validate(ctx context.Context, e bookkeeping.Block) (
 	}
 
 	return validatedBlock{vb: lvb}, nil
-}
-
-type blockFactory struct {
-	*pools.TransactionPool
-}
-
-// AssembleBlock implements Ledger.AssembleBlock.
-func (i blockFactory) AssembleBlock(round basics.Round, deadline time.Time) (agreement.ValidatedBlock, error) {
-	lvb, err := i.TransactionPool.AssembleBlock(round, deadline)
-	if err != nil {
-		return nil, err
-	}
-	return validatedBlock{vb: lvb}, nil
-}
-
-// validatedBlock satisfies agreement.ValidatedBlock
-type validatedBlock struct {
-	vb *ledger.ValidatedBlock
-}
-
-// WithSeed satisfies the agreement.ValidatedBlock interface.
-func (vb validatedBlock) WithSeed(s committee.Seed) agreement.ValidatedBlock {
-	lvb := vb.vb.WithSeed(s)
-	return validatedBlock{vb: &lvb}
-}
-
-// Block satisfies the agreement.ValidatedBlock interface.
-func (vb validatedBlock) Block() bookkeeping.Block {
-	blk := vb.vb.Block()
-	return blk
 }
 
 // agreementLedger implements the agreement.Ledger interface.
